backend/requirements/base.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
redis==6.4.0
pydantic-settings==2.11.0
python-multipart==0.0.20
pyjwt==2.10.1
jinja2==3.1.6
